feat(dockerfile): added dockerfile and and added 'd' suffix to binary name
fix #22
